Refactor copy utilities into a helper class
